Refactored project structure for modularity. Created booking_scheduling app with endpoints to handle service booking requests.
