+++ app/core/admin.py
@@ -49,3 +49,4 @@ admin.site.register(models.UserProfile)
 admin.site.register(models.ProviderProfile)
 admin.site.register(models.ServiceType)
 admin.site.register(models.Review)
+admin.site.register(models.ServiceOrder)
