+++ app/service/urls.py
@@ -1,18 +1,15 @@
-"""
-URL Mappings for the Provider app.
-"""
 from django.urls import (
     path,
     include,
 )
 
 from rest_framework.routers import DefaultRouter
-from provider import views
+from service import views
 
-app_name = 'provider'
+app_name = 'service'
 
 router = DefaultRouter()
-router.register('profile', views.ProviderViewSet, basename='profile')
+router.register('service-type', views.ServiceTypeViewSet, basename='service')
 
 urlpatterns = [
     path('', include(router.urls)),
