+++ app/user/urls.py
@@ -4,27 +4,10 @@ URL mappings for the user API
 from django.urls import path
 from user import views
 
-from rest_framework_simplejwt.views import (
-    TokenObtainPairView,
-    TokenRefreshView,
-    TokenVerifyView,
-)
-
 
 app_name = 'user'
 
 urlpatterns = [
     path('create/', views.CreateUserView.as_view(), name='create'),
     path('me/', views.ManageUserView.as_view(), name='me'),
-
-    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
-    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
-    path('token/verify/', TokenVerifyView.as_view(), name='token_verify'),
-
-    path(
-        'user-profile/',
-        views.ManageUserProfileView.as_view(),
-        name='manage_user_profile'
-    )
-
 ]
