+++ app/user_profile/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class UserProfileConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'user_profile'
